refactor: extract year card component
